# Improve MarshalBinary and UnmarshalBinary performance