- Makes `MarshalBinary` an alias to `ToBytes`
- Removes unnecessary intermediate `bufio.Reader` from `UnmarshalBinary`
```
name               old time/op    new time/op    delta
MarshalBinary-4       125µs ± 3%      69µs ± 0%  -44.26%  (p=0.008 n=5+5)
UnmarshalBinary-4     100µs ± 0%      58µs ± 3%  -41.85%  (p=0.016 n=4+5)

name               old alloc/op   new alloc/op   delta
MarshalBinary-4       509kB ± 0%     276kB ± 0%  -45.73%  (p=0.079 n=4+5)
UnmarshalBinary-4     291kB ± 0%     155kB ± 0%  -46.60%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
MarshalBinary-4        10.0 ± 0%       9.0 ± 0%  -10.00%  (p=0.008 n=5+5)
UnmarshalBinary-4       340 ± 0%       337 ± 0%   -0.88%  (p=0.008 n=5+5)
```