@lemire 2019-05-27
Looks good.